Hydra.Network.Ports

Description

Utility module providing functions to find and allocate random ports.

Documentation

getRandomPort :: IO PortNumber

Find a TCPv4 port which is likely to be free for listening on localhost. This binds a socket, receives an OS-assigned port, then closes the socket.

Note that this is vulnerable to race conditions if another process binds the port returned by getRandomPort before this process does.

Do not use this unless you have no other option.

withFreePort :: (Int -> IO ()) -> IO ()

Find a free TCPv4 port and pass it to the given action.

Should be used only for testing, see getRandomPort for limitations.

isPortOpen :: SockAddr -> IO Bool

Checks whether connect() to a given TCPv4 SockAddr succeeds or returns eCONNREFUSED.

Rethrows connection exceptions in all other cases (e.g. when the host is unroutable).

Code courtesy of nh2: https://stackoverflow.com/a/57022572

unsafePortNumber :: SockAddr -> PortNumber

Get the underlying port number for the given socket address. Fails for UNIX socket addresses which aren't bound to any TCP port.

simpleSockAddr :: (Word8, Word8, Word8, Word8) -> PortNumber -> SockAddr

Creates a SockAttr from host IP and port number.

Example: > simpleSockAddr (127,0,0,1) 8000

randomUnusedTCPPorts :: Int -> IO [Int]

Get a list of random TCPv4 ports that currently do not have any servers listening on them. It may return less than the requested number of ports.

Note that this method of allocating ports is subject to race conditions. Production code should use better methods such as passing a listening socket to the child process.